+++ src/main/php/lang/ast/Emitter.class.php
@@ -233,7 +233,13 @@ abstract class Emitter {
     static $native= ['string' => true, 'int' => true, 'float' => true, 'bool' => true, 'array' => true, 'object' => true];
 
     $name= $node->value[0]->name();
-    if (isset($native[$name])) {
+    if ('?' === $name{0}) {
+      $t= $this->temp();
+      $this->out->write('(null === ('.$t.'= ');
+      $this->emit($node->value[1]);
+      $this->out->write(') ? null : ');
+      $this->out->write(isset($native[$name]) ? '('.$node->value[0]->literal().')'.$t.')' : 'cast('.$t.', \''.substr($name, 1).'\'))');
+    } else if (isset($native[$name])) {
       $this->out->write('('.$node->value[0]->literal().')');
       $this->emit($node->value[1]);
     } else {